Update diskspace_telegram_notification.py
The test free <= freeSpaceTreshold (where free>freeSpaceTreshold) had different results on different machines. I can't explain why in the world it gave True on some machines and False on others, but converting the numbers to float seems to fix it. All machines had the same python3 version installed.

Additionally I added the possibility to define a hermes node and to catch a invalid Nodetype.